ENG-0000 - Support Primary/Inert Activation Modes
Adds a property to route definitions that allows the schema to indicate
which instance of a given route should be activated when there are
multiple possible matches.

In practice, this satisfies a need created by the Notifications/v4
feature, which has multiple routes referring to the same locations in
console.account.