ENG-28315 - Normalize XHRHttpRequest.withCredentials
The only API call in the entire IWS stack is the AIMS authenticate call
-- no other API relies on cookies or `Authorization` headers for any of
their work.  Unfortunately, yarp allows Access-Control-Allow-Credentials
for all endpoints, which is *technically* a vulnerability (Red Siege
identified it as a Low Criticality vulnerability, and Platform Services
and I agree it doesn't constite a real active threat).  To get it off of
our audit results, we will update the UI to use
`Access-Control-Allow-Credentials: true` ONLY for authentication
requests, and `Access-Control-Allow-Credentials: false` for all other
requests.  After the UI is fully updated, yarp will be updated to
disable credentials for all but the authentication endpoint.

Updated version to 1.0.136.