ENG-0000/VTT-4761 - MFA Enrollment Behavior
For some reason, when user management and MFA enrollment were shifted
from console.account to MAGMA, a different behavior emerged when
enrolling an MFA device from an unauthenticated state: the enrollment
call tries to resolve the location of AIMS via the endpoint service and
fails (because there is no authentication token), and instead of falling
back and using the defaults, it simply cancels the action.  A second
call to the enrollment method used the cached default value, but it did
require a second attempt.

This simply disables endpoint resolution for the MFA enrollment methods.