ENG-0000 - Fix Location Resolution
So that it works as it did before the refactor.  The location matrix
(AlLocatorMatrix, now a child of AlExecutionContext) maintains the
current state of which datacenter the user is currently working with,
and correlates defender and insight location identifiers.  The refactor
inadvertantly caused this state not to be persisted as expected.

Appropriately, the region of code where this work happens is labelled
with a warning comment (e.g., "watch out for black magic, kids") from the distant past.